fix(props-docs): fixing Modal, Drawer, Image, and Show components props-docs json generation
