+++ packages/image/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./image"
+export * from "./img"
 export * from "./use-image"
