+++ packages/media-query/src/index.ts
@@ -3,3 +3,5 @@ export * from "./media-query.hook"
 export * from "./use-breakpoint"
 export * from "./use-media-query"
 export * from "./use-breakpoint-value"
+export * from "./show"
+export * from "./hide"
