+++ packages/modal/src/modal.tsx
@@ -1,27 +1,12 @@
-import { CloseButton, CloseButtonProps } from "@chakra-ui/close-button"
-import { FocusLock, FocusLockProps } from "@chakra-ui/focus-lock"
+import { FocusLockProps } from "@chakra-ui/focus-lock"
 import { Portal, PortalProps } from "@chakra-ui/portal"
 import { createContext } from "@chakra-ui/react-context"
 import {
-  chakra,
-  ChakraProps,
-  forwardRef,
-  HTMLChakraProps,
   SystemStyleObject,
   ThemingProps,
   useMultiStyleConfig,
 } from "@chakra-ui/system"
-import { fadeConfig } from "@chakra-ui/transition"
-import { callAllHandlers, cx } from "@chakra-ui/shared-utils"
-import {
-  AnimatePresence,
-  HTMLMotionProps,
-  motion,
-  usePresence,
-} from "framer-motion"
-import { useEffect } from "react"
-import { RemoveScroll } from "react-remove-scroll"
-import { ModalTransition } from "./modal-transition"
+import { AnimatePresence } from "framer-motion"
 import { useModal, UseModalProps, UseModalReturn } from "./use-modal"
 
 interface FocusableElement {
@@ -204,321 +189,3 @@ Modal.defaultProps = {
 }
 
 Modal.displayName = "Modal"
-
-export interface ModalContentProps extends HTMLChakraProps<"section"> {
-  /**
-   * The props to forward to the modal's content wrapper
-   */
-  containerProps?: HTMLChakraProps<"div">
-}
-
-const MotionDiv = chakra(motion.div)
-
-/**
- * ModalContent is used to group modal's content. It has all the
- * necessary `aria-*` properties to indicate that it is a modal
- */
-export const ModalContent = forwardRef<ModalContentProps, "section">(
-  (props, ref) => {
-    const { className, children, containerProps: rootProps, ...rest } = props
-
-    const { getDialogProps, getDialogContainerProps } = useModalContext()
-
-    const dialogProps = getDialogProps(rest, ref) as any
-    const containerProps = getDialogContainerProps(rootProps)
-
-    const _className = cx("chakra-modal__content", className)
-
-    const styles = useModalStyles()
-
-    const dialogStyles: SystemStyleObject = {
-      display: "flex",
-      flexDirection: "column",
-      position: "relative",
-      width: "100%",
-      outline: 0,
-      ...styles.dialog,
-    }
-
-    const dialogContainerStyles: SystemStyleObject = {
-      display: "flex",
-      width: "100vw",
-      height: "100vh",
-      "@supports(height: -webkit-fill-available)": {
-        height: "-webkit-fill-available",
-      },
-      position: "fixed",
-      left: 0,
-      top: 0,
-      ...styles.dialogContainer,
-    }
-
-    const { motionPreset } = useModalContext()
-
-    return (
-      <ModalFocusScope>
-        <chakra.div
-          {...containerProps}
-          className="chakra-modal__content-container"
-          // tabindex="-1" means that the element is not reachable via sequential keyboard navigation, @see #4686
-          tabIndex={-1}
-          __css={dialogContainerStyles}
-        >
-          <ModalTransition
-            preset={motionPreset}
-            className={_className}
-            {...dialogProps}
-            __css={dialogStyles}
-          >
-            {children}
-          </ModalTransition>
-        </chakra.div>
-      </ModalFocusScope>
-    )
-  },
-)
-
-ModalContent.displayName = "ModalContent"
-
-interface ModalFocusScopeProps {
-  /**
-   * @type React.ReactElement
-   */
-  children: React.ReactElement
-}
-
-export function ModalFocusScope(props: ModalFocusScopeProps) {
-  const {
-    autoFocus,
-    trapFocus,
-    dialogRef,
-    initialFocusRef,
-    blockScrollOnMount,
-    allowPinchZoom,
-    finalFocusRef,
-    returnFocusOnClose,
-    preserveScrollBarGap,
-    lockFocusAcrossFrames,
-  } = useModalContext()
-
-  const [isPresent, safeToRemove] = usePresence()
-
-  useEffect(() => {
-    if (!isPresent && safeToRemove) {
-      setTimeout(safeToRemove)
-    }
-  }, [isPresent, safeToRemove])
-
-  return (
-    <FocusLock
-      autoFocus={autoFocus}
-      isDisabled={!trapFocus}
-      initialFocusRef={initialFocusRef}
-      finalFocusRef={finalFocusRef}
-      restoreFocus={returnFocusOnClose}
-      contentRef={dialogRef}
-      lockFocusAcrossFrames={lockFocusAcrossFrames}
-    >
-      <RemoveScroll
-        removeScrollBar={!preserveScrollBarGap}
-        allowPinchZoom={allowPinchZoom}
-        enabled={blockScrollOnMount}
-        forwardProps
-      >
-        {props.children}
-      </RemoveScroll>
-    </FocusLock>
-  )
-}
-
-export interface ModalOverlayProps
-  extends Omit<HTMLMotionProps<"div">, "color" | "transition">,
-    ChakraProps {
-  children?: React.ReactNode
-}
-
-/**
- * ModalOverlay renders a backdrop behind the modal. It is
- * also used as a wrapper for the modal content for better positioning.
- *
- * @see Docs https://chakra-ui.com/modal
- */
-export const ModalOverlay = forwardRef<ModalOverlayProps, "div">(
-  (props, ref) => {
-    const { className, transition, ...rest } = props
-    const _className = cx("chakra-modal__overlay", className)
-
-    const styles = useModalStyles()
-    const overlayStyle: SystemStyleObject = {
-      pos: "fixed",
-      left: "0",
-      top: "0",
-      w: "100vw",
-      h: "100vh",
-      ...styles.overlay,
-    }
-
-    const { motionPreset } = useModalContext()
-    const motionProps: any = motionPreset === "none" ? {} : fadeConfig
-
-    return (
-      <MotionDiv
-        {...motionProps}
-        __css={overlayStyle}
-        ref={ref}
-        className={_className}
-        {...rest}
-      />
-    )
-  },
-)
-
-ModalOverlay.displayName = "ModalOverlay"
-
-export interface ModalHeaderProps extends HTMLChakraProps<"header"> {}
-
-/**
- * ModalHeader
- *
- * React component that houses the title of the modal.
- *
- * @see Docs https://chakra-ui.com/modal
- */
-export const ModalHeader = forwardRef<ModalHeaderProps, "header">(
-  (props, ref) => {
-    const { className, ...rest } = props
-
-    const { headerId, setHeaderMounted } = useModalContext()
-
-    /**
-     * Notify us if this component was rendered or used,
-     * so we can append `aria-labelledby` automatically
-     */
-    useEffect(() => {
-      setHeaderMounted(true)
-      return () => setHeaderMounted(false)
-    }, [setHeaderMounted])
-
-    const _className = cx("chakra-modal__header", className)
-
-    const styles = useModalStyles()
-    const headerStyles: SystemStyleObject = {
-      flex: 0,
-      ...styles.header,
-    }
-
-    return (
-      <chakra.header
-        ref={ref}
-        className={_className}
-        id={headerId}
-        {...rest}
-        __css={headerStyles}
-      />
-    )
-  },
-)
-
-ModalHeader.displayName = "ModalHeader"
-
-export interface ModalBodyProps extends HTMLChakraProps<"div"> {}
-
-/**
- * ModalBody
- *
- * React component that houses the main content of the modal.
- *
- * @see Docs https://chakra-ui.com/modal
- */
-export const ModalBody = forwardRef<ModalBodyProps, "div">((props, ref) => {
-  const { className, ...rest } = props
-  const { bodyId, setBodyMounted } = useModalContext()
-
-  /**
-   * Notify us if this component was rendered or used,
-   * so we can append `aria-describedby` automatically
-   */
-  useEffect(() => {
-    setBodyMounted(true)
-    return () => setBodyMounted(false)
-  }, [setBodyMounted])
-
-  const _className = cx("chakra-modal__body", className)
-  const styles = useModalStyles()
-
-  return (
-    <chakra.div
-      ref={ref}
-      className={_className}
-      id={bodyId}
-      {...rest}
-      __css={styles.body}
-    />
-  )
-})
-
-ModalBody.displayName = "ModalBody"
-
-export interface ModalFooterProps extends HTMLChakraProps<"footer"> {}
-
-/**
- * ModalFooter houses the action buttons of the modal.
- * @see Docs https://chakra-ui.com/modal
- */
-export const ModalFooter = forwardRef<ModalFooterProps, "footer">(
-  (props, ref) => {
-    const { className, ...rest } = props
-    const _className = cx("chakra-modal__footer", className)
-
-    const styles = useModalStyles()
-    const footerStyles: SystemStyleObject = {
-      display: "flex",
-      alignItems: "center",
-      justifyContent: "flex-end",
-      ...styles.footer,
-    }
-
-    return (
-      <chakra.footer
-        ref={ref}
-        {...rest}
-        __css={footerStyles}
-        className={_className}
-      />
-    )
-  },
-)
-
-ModalFooter.displayName = "ModalFooter"
-
-/**
- * ModalCloseButton is used closes the modal.
- *
- * You don't need to pass the `onClick` to it, it reads the
- * `onClose` action from the modal context.
- */
-export const ModalCloseButton = forwardRef<CloseButtonProps, "button">(
-  (props, ref) => {
-    const { onClick, className, ...rest } = props
-    const { onClose } = useModalContext()
-
-    const _className = cx("chakra-modal__close-btn", className)
-
-    const styles = useModalStyles()
-
-    return (
-      <CloseButton
-        ref={ref}
-        __css={styles.closeButton}
-        className={_className}
-        onClick={callAllHandlers(onClick, (event: React.MouseEvent) => {
-          event.stopPropagation()
-          onClose()
-        })}
-        {...rest}
-      />
-    )
-  },
-)
-
-ModalCloseButton.displayName = "ModalCloseButton"
